requirements.txt
-i https://pypi.org/simple
asgiref==3.8.1; python_version >= '3.8'
django==4.2.8; python_version >= '3.8'
django-tastypie==0.14.6
gunicorn==21.2.0; python_version >= '3.5'
packaging==24.0; python_version >= '3.7'
psycopg2==2.9.9; python_version >= '3.7'
python-dateutil==2.9.0.post0; python_version >= '2.7' and python_version not in '3.0, 3.1, 3.2, 3.3'
python-mimeparse==1.6.0
six==1.16.0; python_version >= '2.7' and python_version not in '3.0, 3.1, 3.2, 3.3'
sqlparse==0.4.4; python_version >= '3.5'
tzdata==2024.1; sys_platform == 'win32'
whitenoise==6.6.0; python_version >= '3.8'
